Add Queue page and update AppBar, AppMenu, and CSS imports
This commit introduces a new Queue page that includes hooks, state management, and functions to handle fetching, updating, and enqueuing jobs. It also updates App.jsx with CSS imports for correct rendering of ag-grid in Queue page. The navigation and arrangement of the application were enhanced by adding links to the new Queue page in AppBar and AppMenu components. Lastly, ag-grid-react library has been added to the application for data grid functionality in the Queue page.